// src/lib2/models/Workspace.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Workspace = {
    color?: string;
    created_at?: string;
    description?: string;
    id?: string;
    name?: string;
    org_id?: string;
    pipelines_count?: number;
};